# Conflicts:
#	src/main/java/cnpr/lcss/dao/Notification.java
#	src/main/java/cnpr/lcss/service/NotificationService.java

+++ src/main/java/cnpr/lcss/model/FeedbackDto.java
@@ -13,7 +13,7 @@ public class FeedbackDto {
     private String studentUsername;
     private String studentName;
     private String studentImage;
-    private int teacherRating;
-    private int subjectRating;
+    private double teacherRating;
+    private double subjectRating;
     private String feedback;
 }
